requirements-setup.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements-setup.txt requirements-setup.in
#
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via pytest
boto3==1.4.7
botocore==1.7.36          # via boto3, s3transfer
click==6.7                # via pip-tools, python-dotenv
coverage==4.5.1
docutils==0.14            # via botocore
faker==0.9.0
first==2.0.1              # via pip-tools
funcsigs==1.0.2           # via mock, pytest
futures==3.1.1            # via s3transfer
invoke==0.21.0
ipaddress==1.0.22         # via faker
jmespath==0.9.3           # via boto3, botocore
mock==2.0.0               # via pytest-mock
more-itertools==4.3.0     # via pytest
pathlib2==2.3.2           # via pytest
pbr==4.2.0                # via mock
pip-tools==2.0.2
placebo==0.8.2
pluggy==0.7.1             # via pytest
py==1.5.4                 # via pytest
pytest-cov==2.5.1
pytest-html==1.19.0
pytest-metadata==1.7.0    # via pytest-html
pytest-mock==1.10.0
pytest==3.7.1
python-dateutil==2.6.1
python-dotenv==0.7.1
pyyaml==3.13
s3transfer==0.1.11        # via boto3
scandir==1.8              # via pathlib2
shortuuid==0.5.0
six==1.11.0               # via faker, mock, more-itertools, pathlib2, pip-tools, pytest, python-dateutil
text-unidecode==1.2       # via faker
wheel==0.31.1

# The following packages are considered to be unsafe in a requirements file:
# setuptools==44.1.0        # via pytest
